requirements.txt
# Existing dependencies
requests==2.32.5
python-dotenv==1.2.1
openai==2.14.0

# Load Balancer Server dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
aiosqlite==0.20.0
tabulate==0.9.0
pydantic>=2.12.0
